backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
# psycopg2-binary==2.9.10 # Commented out as we use SQLite and it fails on some Windows setups
pydantic==2.10.3
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.17
aiofiles==24.1.0
reportlab==4.2.5
python-docx==1.1.2
openpyxl==3.1.5
pandas==2.2.3
numpy==2.2.1
scikit-learn==1.6.0
pydantic-core==2.27.1
pymysql==1.1.1
cryptography==44.0.0
anthropic==0.40.0
pdfplumber==0.11.4
python-dotenv==1.0.1
starlette==0.41.2
email-validator==2.2.0
google-generativeai==0.8.3
groq>=0.11.0
pypdf>=4.0.0
